requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlmodel>=0.0.21
anthropic>=0.40
feedparser>=6.0
httpx>=0.27
python-dotenv>=1.0
rich>=13.0
python-dateutil>=2.9
beautifulsoup4>=4.12
lxml>=5.0
